requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Alembic==2.0.1
Flask-Assets==2.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.3
Flask-Uploads==0.2.1
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.2
Pillow==7.1.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
visitor==0.1.3
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.1
